Started to work on the JADE code. WIP HostCache behaviour.
Receiving the request for neighbours behaviour is complete, todo: reply
with a list of neighbours.  Going well, but concerning that it will be
some time with lots of code written before this can be executed.

I'm worried about an error thrown up in trying to match against the
message.  There was a discrepency between two types of ACLMessage - one
under an 'interrogation' package and one under something more general.
What if it is not possible to retrieve the contents of a message in the
Matcher class?  If so, I'm not sure if it's even possible to match a
message based on its Action field in the ontology.  I'm surprised how
difficult it is to filter messages dynamically - i.e., to define your
own set of message identifiers, and then create a MessageTemplate which
only matches a given identifier.